soon

+++ Predicate.cpp
@@ -11,7 +11,7 @@ void Predicate::addParameter(std::vector<Parameter*> input) {
 void Predicate::stringPredicate(std::vector<Predicate> input) {
     for (auto & i : input) {
         std::cout << "  " << i.id << "(";
-        for (int j = 0; j < i.parameters.size(); j++) {
+        for (long unsigned int j = 0; j < i.parameters.size(); j++) {
             if (j == i.parameters.size() - 1) {
                 std::cout << i.parameters.at(j)->paramString() << ")";
             }
