+++ Rule.cpp
@@ -12,7 +12,7 @@ void Rule::setHead(const Predicate& input) {
 void Rule::ruleString(std::vector<Rule> input) {
     for (auto & i : input) {
         std::cout << "  " << i.headPredicate.id << "(";
-        for (int j = 0; j < i.headPredicate.parameters.size(); j++) {
+        for (long unsigned int j = 0; j < i.headPredicate.parameters.size(); j++) {
             if (j == i.headPredicate.parameters.size()-1) {
                 std::cout << i.headPredicate.parameters.at(j)->paramString() << ")";
             }
@@ -23,7 +23,7 @@ void Rule::ruleString(std::vector<Rule> input) {
         std::cout << " :- ";
         for (auto & bodyPredicate : i.bodyPredicates) {
             std::cout << bodyPredicate.id << "(";
-            for (int h = 0; h < bodyPredicate.parameters.size(); h++) {
+            for (long unsigned int h = 0; h < bodyPredicate.parameters.size(); h++) {
                 if (h == bodyPredicate.parameters.size()-1){
                     std::cout << bodyPredicate.parameters.at(h)->paramString() << ")";
                 }
